+++ Scripts/Data/ButtonEvent.cs
@@ -1,0 +1,22 @@
+﻿/*
+ USTRAP
+ DATA CLASS
+ BUTTON EVENT
+ v1.1
+ LAST EDITED: TUESDAY NOVEMBER 29, 2022
+ COPYRIGHT © TECH SKULL STUDIOS
+*/
+
+using System;
+using UnityEngine.UIElements;
+
+namespace UStrap.Data
+{
+    /// <summary>
+    /// Wrapper for button events in the inspector.
+    /// </summary>
+    [Serializable]
+    public class ButtonEvent : UIElementClickEvent<Button>
+    { 
+    }
+}