+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+[Version 0.3 | December 19, 2022]
+• Created new UIElemementValueChangedEvent abstract class in Data
+• Created new TextFieldEvent wrapper class in Data
+• Created new ToggleEvent wrapper class in Data
+• Created new LabelEvent wrapper class in Data
+• Created new SliderEventMinMax wrapper class in Data
+• Created new SliderEventInt wrapper class in Data
+• Created new ScrollerEvent wrapper class in Data
+• Created new FoldoutEvent wrapper class in Data
+• Created new IQuery interface in Ustrap
+• Modified the DOM Manager to include events for every type of VisualElement
+
 [Version 0.2 | November 29, 2022]
 • Created Unity package for UStrap
 • Updating USS style to use Bootstrap 5